requirements.txt
torch==2.6.0+cu124
torchvision==0.21.0+cu124
timm==1.0.22
transformers==4.57.1
accelerate==1.9.0
numpy==1.26.4
pandas==2.3.1
scikit-learn==1.8.0
matplotlib==3.10.3
seaborn==0.13.2
tqdm==4.67.1
pillow==11.0.0
huggingface_hub==0.29.0
swanlab==0.7.6